Build data/publication/index.tsv
We've not being doing a great job of maintaining this file,
and it's a source of conflicts. It's primarily to help loading
data from the built files via HTTP, which we've started to move
to S3/digital-land-data repository, so let's make it a build
artefact rather than something we manually edit.